requirements.txt
1.tensorflow-cpu
2.opencv-python-headless
3.numpy
4.pillow